Switch from decimal to double for non-financial app. Replaced `DebuggerDisplayAttribute` with overridden ToString and tested public method.

+++ Breadmaker/Ingredient.cs
@@ -1,3 +1,3 @@
 namespace Breadmaker;
 
-public record Ingredient(string Name, decimal Mass);
+public record Ingredient(string Name, double Mass);
